+++ public/locale/en.json
@@ -937,6 +937,7 @@
   "file_preview_not_supported": "Can't preview this file. Try downloading it.",
   "file_type": "File Type",
   "file_uploaded": "File Uploaded Successfully",
+  "files": "Files",
   "filter": "Filter",
   "filter_by": "Filter By",
   "filter_by_category": "Filter by category",
