+++ src/components/Patient/PatientDetailsTab/PatientFiles.tsx
@@ -1,0 +1,13 @@
+import { FilesTab } from "@/components/Files/FilesTab";
+
+import { PatientProps } from ".";
+
+export const PatientFilesTab = (props: PatientProps) => {
+  return (
+    <FilesTab
+      type="patient"
+      facilityId={props.facilityId}
+      patientId={props.patientData.id}
+    />
+  );
+};
